+++ NEWS.md
@@ -12,6 +12,8 @@
 - The results of `cohens_d()` is now a data frame containing the Cohen's d and the magnitude.
 
 ## Bug fixes
+  
+- the argument `detatiled` is now passed to `compare_pairs()`.
 
 # rstatix 0.1.0