+++ NEWS.md
@@ -4,14 +4,18 @@
    
 ## New features
   
+- New function `get_pvalue_position` added to autocompute p-value positions for plotting significance using ggplot2.
+- New function `get_comparisons()` added to create a list of possible pairwise comparisons between groups.
 - New function `dunn_test()` added for multiple pairwise comparisons following Kruskal-Wallis test.
 - New function `sign_test()` added.
+
 
 ## Major changes
    
    
 ## Minor changes
    
+- `get_summary_stats()` now supports type = "min", "max", "mean" or "median"
 - the results of `t_test()`, `wilcox_test()`, `dunn_test()` and `sign_test()` are now an object of class `rstatix_test` that has an attribute named **args** for holding the test arguments.
 - The results of `cohens_d()` is now a data frame containing the Cohen's d and the magnitude.